requirements.txt
numpy
openai
pandas
python-dotenv
requests
vaderSentiment
selenium